pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "uitag"
version = "0.2.3"
description = "SoM (Set-of-Mark) detection pipeline for macOS — Apple Vision + Florence-2 on MLX"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{name = "Swaylen Hayes"}]
keywords = ["uitag", "som", "set-of-mark", "detection", "apple-vision", "mlx", "florence-2", "ui-detection"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "mlx-vlm>=0.3.12",
    "pillow>=10.0.0",
    "transformers>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-benchmark>=4.0.0",
    "ruff>=0.1.0",
    "jsonschema>=4.0.0",
]
coreml = [
    "coremltools>=8.0",
    "torch>=2.0",
]

[project.urls]
Homepage = "https://github.com/swaylenhayes/uitag"
Repository = "https://github.com/swaylenhayes/uitag"
Issues = "https://github.com/swaylenhayes/uitag/issues"

[project.scripts]
uitag = "uitag.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["uitag*"]

[tool.setuptools.package-data]
uitag = ["tools/*.swift", "*.json"]

[tool.ruff]
line-length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
